Added hooks and instance method to customer model. Added utilities folder to clean up controllers
