--- server/resources/static/assets/uploads/1630437233492-customers.csv
@@ -1,51 +1,0 @@
-Email,First Name,Last Name,Phone number,City,State,Last Order Price,Last Order Date
-"spa,dkins@att.net",Giovanny,Atkinson,(401)433-3114,Warner,NH,35,8/12/2021
-zilla@optonline.net,Olive,Lang,(618)472-3603,East Natchitoches,,120,8/16/2021
-harryh@icloud.com,Santiago,Riddle,(425)423-4138,Lyon,WV,12,8/26/2021
-marslo@aol.co,Marcel,Sloan,(616)889-7721,Willow Run,IL,89.99,9/9/2021
-wkrebs@sbcglobal.net,Drake,Rebs,(219)683-8767,Conyersville,AZ,44,9/17/2021
-sopwith@aol.com,James,Randolph,(507)623-3577,Mount Baker,NY,45,9/23/2021
-gavinls@sbcglobal.net,Gavin,Obrien,(786)441-2729,Farmington Lake,OK,95,9/30/2021
-lstein@a  ol.com,Owen,Stein,(769)355-5422,Martins Corner,TX,32,10/6/2021
-dgatwood@optonline.net,Nelson,Silva,(303)961-5760,Pickerel Narrows,MT,13.99,10/15/2021
-"danzigis,m@icloud.com",Dan,Zigi,(256)931-9696,Willaha,OH,12,10/21/2021
-bcevc@icloud.com,Rodrigo,Cuevas,(308)204-9038,Center,MA,11,10/26/2021
-frederic@sbcglobal.net,Frederic,Lee,(205)365-8292,Spring City,MS,87,10/27/2021
-bockelboy  @me.com,Joyce,Frank,(319)635-4374,Mittenlane,TX,49.99,10/28/2021
-pereinar@sbcglobal.net,Korbin,Pereinar,(415)572-8494,East Waterford,ME,22,10/29/2021
-rddesign@live.com,Howard,Carrillo,(289)672-8870,Coltman,WV,12,11/3/2021
-"naoya@msn.co,",Nayomi,Dawson,(276)335-6240,Scottsville,,12,11/9/2021
-jesse@sbcglobal.net,Jesse,Hahn,(231)823-9734,Hebron,AZ,49.99,11/10/2021
-syrinx@optonline.net,Meadow,Berger,(670)233-6224,Longview,MA,85,11/22/2021
-presoff@ms'n.com,Whitney,Shelton,(334)201-6824,Emerson,MT,116,12/1/2021
-bachmann@hotmail.com,Paris,Mann,(336)283-4614,North Knoxville,AL,16,12/3/2021
-joelw@yahoo.ca,Joel,Schaefer,(336)877-9176,Ipswich,IA,11,12/7/2021
-joelw@yahoo.ca,Joel,Schaefer,(336)877-9176,Ipswich,IA,76.99,12/13/2021
-fviegas@mac.com,Rafael,Viegas,(214)969-3080,Storms,TX,39,12/22/2021
-trieuvan@aol.com,Santino,Cabrera,(747)900-6516,Kalauao,SD,87,12/24/2021
-storerm@sbcglobal.net,Nayeli,Juarez,(657)378-3335,Farwell,FL,33,12/30/2021
-cderoove@gmail.com,Zaiden,Duran,(618)231-7450,Brentwood Village,MT,45,8/10/2021
-intlprog@hotmail.com,Jamie,Martinez,(617)889-8674,Wilhelm Park,MT,54,8/12/2021
-jaffe...@verizon.net,Charlize,Perez,(289)572-1200,Bannister Acres,MC,89.99,8/24/2021
-lpalmer@icloud.com,Sara,Palmer,(217)657-9901,Bent Pine,VA,23,8/25/2021
-ahwon@icloud.com,Alan,Won,(516)992-9423,Mitchell,AZ,33,8/30/2021
-mrobshaw@outlook.com,Rob,Shaw,(615)221-5554,Social Circle,MO,54,9/8/2021
-a~dhere@aol.com,Fisher,Here,(434)299-8652,Kreutzberg,NY,76.99,9/17/2021
-malattia@yahoo.ca,Lucille,Malattia,(608)380-4822,Cimarron,WA,15,9/23/2021
-jesse@sbcglobal.net,Jesse,Hahn,(231)823-9734,Brookhaven,,11,9/24/2021
-errxn@sbcglobal.net,Jordyn,Mullen,(385)207-5077,Montverde Junction,NJ,28,9/29/2021
-reziac@gmail.com,Marshall,Landry,(551)587-7297,Midland City,AK,28.45,9/30/2021
-bonmots@comcast.net,Jaslene,Whitaker,(352)434-1651,Sacramento,ME,79.99,10/8/2021
-munson@live.com,Arturo,Cisneros,(347)563-4789,Del Rey Oaks,RI,120,10/11/2021
-rogerspl@mac.com,Roger,Brown,(323)298-3567,Coal Creek,OK,13.99,10/20/2021
-mbrown@live.com,Shaniya,Brown,(208)255-2382,Rabbitown,TN,26,11/10/2021
-wojciech@o..utlook.com,Rhianna,Armstrong,(575)628-5858,Fairland,GA,89.99,11/11/2021
-meinkej@yahoo.com,Amelie,Mosley,(667)303-9303,Gaskil,DE,26,11/17/2021
-pgottsch@mac.com,Esperanza,Knapp,(810)234-9251,Morgan Mill,OK,120,11/18/2021
-murdocj@optonline.net,Murdoc,Wilkinson,(806)774-7884,Merrimac South,AL,76.99,11/24/2021
-miyop@verizon.net,Reagan,Sweeney,(216)271-8398,Stanardsville,NH,13.99,11/26/2021
-jtorkbob@comcast.net,Dahlia,Arellano,(618)691-4399,Two Brooks,WI,26,11/30/2021
-louise'@me.com,Louise,Barber,(336)929-8671,Curriers,NM,89.99,12/3/2021
-fukuchi@gmail.com,Jayden,Chi,(305)674-4548,Skookumchuck,VA,26,12/7/2021
-andale@icloud.com,Andres,Ramon,(413)785-5626,Edgerton,RI,76.99,12/20/2021
-webteam@msn.com,Kyan,Harper,(209)828-3279,Slater,MO,13.99,12/23/2021
